Refactor functions from io.clj
Given the upcoming addition of tags, we needed a way to add metadata to posts
using something other than the post's file name. The idea here is that we
will now add metadata to the top of posts in the form of a clojure map.
We also wrote functions to parse this metadata while still being able to
slurp markdown from posts.